Data set combination and manipulation for use in HEI scoring.

combo returns a single data set combining the FPED and NHANES dietary and demographic data sets. Additionally, it allows for specification of an age threshold such that records with age values below this number will be excluded from the output. Furthermore, any rows for which dietary recall was deemed unreliable (DRSTZ != 1) are automatically dropped. Finally, additional necessary HEI inputs are generated from the existing columns, see "Value" section below.

combo(fped, diet, demograph, agethresh = 2)

Arguments

fped
food pattern equivalent database data

diet
dietary data from NHANES database

demograph
demographic data from NHANES database

agethresh
numeric threshold for age of survey participants to be included

Value

Object of class data.frame representing a data set fully prepped for HEI score calculation. All columns from the three input data sets are preserved while a handful of additional derivative columns are included, specifically:

WHOLEFRT: T_F_CITMLB (citrus, melons, and berries) + T_F_OTHER (fruits excluding citrus, melons, and berries)

MONOPOLY: TMFAT (total monounsaturated fatty acids) + TPFAT (total polyunsaturated fatty acids)

ALLMEAT: T_PF_MPS_TOTAL (meat, poultry, seafood, organ meat, and cured meat) + T_PF_EGGS (eggs and egg substitutes) + T_PF_NUTSDS (peanuts, tree nuts, and seeds) + T_PF_SOY (soy products)

SEAPLANT: T_PF_SEAFD_HI (Seafood high in n-3 fatty acids) + T_PF_SEAFD_LOW (Seafood low in n-3 fatty acids) + T_PF_NUTSDS (peanuts, tree nuts, and seeds) + T_PF_SOY + (soy products)

ADDSUGC: T_ADD_SUGARS (added sugars) * 16

SOLFATC: T_SOLID_FATS (fats naturally present in meat, etc.) * 9

MAXALCGR: TKCAL (total calories) / 1000 * 13

EXALCCAL (amount of alcohol above acceptable threshold): (TALCO (total alcohol) - MAXALCGR) * 7

EMPTYCAL10: ADDSUGC + SOLFATC + EXALCCAL