NHANES demographic data retrieval.

get_demo returns an abridged National Health and Nutrition Examination Survey (NHANES) demographic data set for a given year.

get_demo(year)

Arguments

year
year combination of data set to retrieve ... must be one of the following:

"2005/2006"

"2007/2008"

"2009/2010"

"2011/2012"

"2013/2014"

Value

Object of class data.frame containing a modified NHANES demographic data set for a specified year. The complete NHANES demographic data set contains several more variables than are necessary for computing HEI scores. In fact, the only variable contained in the returned data set (in addition to an NHANES unique sequence identifier for each participant) is not, strictly speaking, even required itself, but becomes relevant if data sets from multiple separate NHANES iterations are concatenated, specifically:

SDDSRVYR: This variable represents the two-year data release cycle number (e.g. a value of “6” denotes NHANES 2009–2010).

References

https://www.cdc.gov/nchs/nhanes/nhanes_questionnaires.htm

Examples

# NOT RUN {
get_demo("2009/2010")
# }