FPED data retrieval.

get_fped returns an abridged Food Patterns Equivalents Database (FPED) data set for a given year.

get_fped(year, day)

Arguments

year
year combination of data set to retrieve ... must be one of the following:

"2005/2006"

"2007/2008"

"2009/2010"

"2011/2012"

"2013/2014"

day
identifier for survey day to retrieve ... must be one of the following:

"first" (data for study participants from first day)

"second" (study participant data from second day)

"both" (data for study participants present in both days, with values averaged across the two days)

Value

Object of class data.frame containing a modified FPED data set for specified year and day. FPED proper contains several more variables than are necessary for computing HEI scores. The data set returned by this function has been trimmed down to contain (in addition to an NHANES unique sequence identifier for each participant, age, and dietary recall status) only the columns of data needed for HEI score calculation, specifically:

T_F_CITMLB: Intact fruits (whole or cut) of citrus, melons, and berries (cup eq.)

T_PF_EGGS: Eggs (chicken, duck, goose, quail) and egg substitutes (oz. eq.)

T_F_OTHER: Intact fruits (whole or cut); excluding citrus, melons, and berries (cup eq.)

T_PF_MPS_TOTAL: Total of meat, poultry, seafood, organ meat, and cured meat (oz. eq.)

T_PF_SOY: Soy products, excluding calcium fortified soy milk (soymilk) and mature soybeans (oz. eq.)

T_PF_SEAFD_HI: Seafood (finfish, shellfish, and other seafood) high in n-3 fatty acids (oz. eq.)

T_PF_SEAFD_LOW: Seafood (finfish, shellfish, and other seafood) low in n-3 fatty acids (oz. eq.)

T_PF_NUTSDS: Peanuts, tree nuts, and seeds; excludes coconut (oz. eq.)

T_SOLID_FATS: Fats naturally present in meat, poultry, eggs, and dairy (lard, tallow, and butter); hydrogenated or partially hydrogenated oils; shortening; palm, palm kernel and coconut oils; fat naturally present in coconut meat and cocoa butter; and 50% of fat present in stick and tub margarines and margarine spreads (grams)

T_ADD_SUGARS: Foods defined as added sugars (tsp. eq.)

T_V_TOTAL: Total dark green, red and orange, starchy, and other vegetables; excludes legumes (cup eq.)

T_V_LEGUMES: Total dark green, red and orange, starchy, and other vegetables; excludes legumes (cup eq.)

T_V_DRKGR: Dark green vegetables (cup eq.)

T_F_TOTAL: Total intact fruits (whole or cut) and fruit juices (cup eq.)

T_G_WHOLE: Grains defined as whole grains and contain the entire grain kernel - the bran, germ, and endosperm (oz. eq.)

T_D_TOTAL: Total milk, yogurt, cheese, and whey. For some foods, the total dairy values could be higher than the sum of D_MILK, D_YOGURT, and D_CHEESE because miscellaneous dairy component composed of whey is not included in FPED as a separate variable. (cup eq.)

T_G_REFINED: Refined grains that do not contain all of the components of the entire grain kernel (oz. eq.)

References

https://www.ars.usda.gov/northeast-area/beltsville-md/beltsville-human-nutrition-research-center/food-surveys-research-group/docs/fped-overview/

Examples

get_fped("2009/2010", "both")

#>       SEQN RIDAGEYR DRSTZ T_F_CITMLB T_PF_EGGS T_F_OTHER T_PF_MPS_TOTAL
#> 1    51624       34     1      0.805     0.150     0.940          2.760
#> 2    51625        4     1      0.000     0.745     0.630          3.850
#> 3    51626       16     1      0.000     1.015     0.000          4.700
#> 4    51627       10     1      0.000     1.085     0.185          4.525
#> 5    51628       60     1      0.000     0.310     0.000          4.025
#> 6    51629       26     1      0.000     0.905     0.000         11.225
#> 7    51630       49     1      0.005     0.000     1.245          8.575
#> 8    51633       80     1      1.650     1.025     0.000          3.945
#> 9    51634       10     1      0.000     0.220     0.000          5.710
#> 10   51635       80     1      0.000     0.025     0.000          5.230
#> 11   51638        9     1      0.000     0.080     0.000          1.420
#> 12   51639        4     1      0.165     1.320     0.165          7.175
#> 13   51640       17     1      0.000     0.185     0.000          9.345
#> 14   51641       13     1      0.000     0.045     0.000          7.460
#> 15   51642        7     1      0.045     0.105     0.485          2.965
#> 16   51643       42     1      0.090     0.225     0.000          1.880
#> 17   51644        0     1      0.000     0.050     0.515          0.000
#> 18   51646        8     1      0.000     0.000     0.390          5.170
#> 19   51647       45     1      0.670     0.020     0.000          2.130
#> 20   51648       28     1      0.000     0.885     0.000          2.900
#> 21   51650       11     1      0.015     0.000     0.535          0.200
#> 22   51651       19     1      0.000     0.045     1.640          2.685
#> 23   51654       66     1      0.000     0.000     1.865          5.195
#> 24   51656       58     1      0.305     0.825     0.700          5.620
#> 25   51657       54     1      0.000     0.660     0.000          2.675
#> 26   51658       26     1      0.705     0.910     0.680          0.755
#> 27   51660       32     1      0.240     0.165     1.440          8.080
#> 28   51661       60     1      0.000     0.820     0.550          6.900
#> 29   51662       44     1      0.000     0.030     0.000          5.445
#> 30   51663       18     1      0.000     1.345     0.040          4.785
#> 31   51664       53     1      0.250     0.070     1.175          3.760
#> 32   51667       50     1      0.260     1.520     2.190          7.915
#> 33   51668       70     1      0.000     0.845     0.910          7.410
#> 34   51670       22     1      0.000     0.000     0.000          5.235
#> 35   51672       27     1      0.000     0.000     0.340          2.595
#> 36   51675       80     1      0.190     0.010     0.270          1.615
#> 37   51676       27     1      0.000     1.155     0.000          6.545
#> 38   51677       33     1      0.165     0.000     0.015         18.690
#> 39   51678       60     1      0.000     0.125     0.185          7.920
#> 40   51679       16     1      0.000     0.020     0.335          0.000
#> 41   51680       60     1      0.000     0.015     0.000          4.355
#> 42   51681        5     1      0.000     0.000     0.000          2.855
#> 43   51682       11     1      0.105     2.330     0.275          2.715
#> 44   51683       35     1      0.000     1.220     0.750          1.510
#> 45   51685       56     1      0.045     0.780     0.680          5.470
#> 46   51687       78     1      1.110     0.110     1.250          3.220
#> 47   51689        8     1      0.000     0.040     0.000          2.920
#> 48   51690       63     1      1.155     0.000     0.000          2.635
#> 49   51691       57     1      0.060     0.000     0.365          1.810
#> 50   51694       38     1      0.000     0.020     0.060          8.890
#>      T_PF_SOY T_PF_SEAFD_HI T_PF_SEAFD_LOW T_PF_NUTSDS T_SOLID_FATS
#> 1       0.000         0.000          0.000       0.000       59.260
#> 2       0.000         0.000          0.000       0.030       22.260
#> 3       0.000         0.000          0.000       0.000       55.295
#> 4       0.000         0.000          0.000       0.000      100.055
#> 5       0.000         0.000          0.000       0.000       30.545
#> 6       0.000         0.000          0.000       0.000      119.660
#> 7       0.000         0.000          0.000       0.335       23.415
#> 8       0.000         0.000          0.000       1.330       18.705
#> 9       0.000         0.000          0.000       0.000       35.680
#> 10      0.000         0.000          0.000       0.000       34.710
#> 11      0.000         0.000          0.000       1.415       42.635
#> 12      0.040         0.000          0.000       0.730       53.275
#> 13      0.000         0.635          2.310       0.000       45.695
#> 14      0.000         0.000          0.000       0.335       25.385
#> 15      0.000         0.000          0.000       0.030       30.490
#> 16      0.000         0.000          0.000       0.140       29.220
#> 17      0.000         0.000          0.000       0.000        0.500
#> 18      0.000         0.000          0.000       0.000       46.925
#> 19      0.000         0.000          1.155       0.000        8.620
#> 20      0.000         0.000          0.000       0.000       16.325
#> 21      0.065         0.000          0.000       0.000       21.240
#> 22      0.040         0.000          0.000       0.000       18.015
#> 23      0.000         0.000          0.000       0.840       23.305
#> 24      0.000         0.000          1.775       0.045       27.415
#> 25      0.000         0.000          0.000       0.485       14.915
#> 26      0.000         0.000          0.000       0.095       21.955
#> 27      0.000         0.000          0.000       0.000       23.300
#> 28      0.000         0.000          0.290       0.000       33.100
#> 29      0.000         0.000          0.390       0.000       17.830
#> 30      0.000         0.000          0.000       0.195       32.615
#> 31      0.000         0.000          0.000       0.455       42.950
#> 32      0.000         0.000          0.000       0.000       57.650
#> 33      0.000         0.000          1.150       0.000       24.730
#> 34      0.000         0.000          0.000       2.735       10.415
#> 35      0.000         0.000          0.000       0.000        3.575
#> 36      0.000         0.000          0.000       0.000       16.810
#> 37      0.000         0.000          0.000       0.000       47.440
#> 38      0.000         0.000          4.710       0.260       39.700
#> 39      0.000         0.000          0.000       0.000       25.075
#> 40      0.000         0.000          0.000       0.000       49.405
#> 41      0.000         0.000          0.000       0.000       14.905
#> 42      0.000         0.000          0.000       0.000       24.015
#> 43      0.000         0.000          0.000       0.205       80.305
#> 44      0.000         0.230          0.710       0.740       64.955
#> 45      0.000         0.000          0.000       1.735       12.105
#> 46      0.020         0.000          0.000       5.475       24.245
#> 47      0.000         0.000          0.000       0.190       48.540
#> 48      0.000         0.000          0.000       0.085       13.500
#> 49      0.000         0.000          0.000       1.440       16.215
#> 50      0.000         0.000          0.000       0.195       56.540
#>      T_ADD_SUGARS T_V_TOTAL T_V_LEGUMES T_V_DRKGR T_F_TOTAL T_G_WHOLE T_D_TOTAL
#> 1          14.850     1.105       0.000     0.000     1.745     0.000     3.050
#> 2          16.685     0.205       0.025     0.000     0.635     1.585     2.325
#> 3          16.270     0.270       0.000     0.000     0.295     3.085     2.490
#> 4          11.590     1.160       0.000     0.000     0.930     2.370     5.780
#> 5          19.515     1.195       0.000     0.685     0.000     0.000     0.185
#> 6          72.590     2.410       0.250     0.180     0.000     1.875     5.985
#> 7           4.720     1.040       0.000     0.000     1.430     1.685     1.170
#> 8           2.470     1.285       0.000     0.030     1.650     0.750     4.655
#> 9          16.805     0.300       0.000     0.000     0.025     0.000     4.620
#> 10          1.090     2.480       0.000     0.000     0.000     1.060     1.555
#> 11         23.425     0.470       0.000     0.000     0.000     0.575     3.705
#> 12         17.260     0.185       0.580     0.000     0.390     1.370     3.105
#> 13         20.560     1.010       0.000     0.000     0.000     0.645     1.300
#> 14         16.850     1.020       0.000     0.000     0.010     0.535     0.080
#> 15          9.275     0.805       0.380     0.035     1.650     0.555     3.810
#> 16         72.220     0.415       0.000     0.000     0.090     0.000     2.685
#> 17          0.495     0.560       0.000     0.000     0.640     0.805     0.195
#> 18         18.555     0.740       0.000     0.000     1.260     0.620     2.980
#> 19          3.105     6.130       0.270     1.380     0.705     0.000     0.895
#> 20         19.555     1.320       1.115     0.000     0.025     0.000     0.235
#> 21          9.950     0.120       0.000     0.000     0.550     0.000     1.420
#> 22         16.595     0.665       0.000     0.000     1.640     2.085     1.485
#> 23         38.880     3.055       0.660     0.685     2.940     4.960     1.115
#> 24         13.500     1.855       0.080     0.065     1.005     2.770     1.310
#> 25          4.185     0.145       0.155     0.000     0.005     0.275     0.365
#> 26         27.705     1.810       0.000     0.010     1.405     0.000     1.360
#> 27         26.260     0.925       0.455     0.060     1.680     0.000     0.650
#> 28         10.510     1.375       0.000     0.110     1.220     0.000     0.415
#> 29          5.325     0.820       0.140     0.000     0.505     0.405     0.865
#> 30         11.740     1.160       0.420     0.000     0.610     1.025     3.075
#> 31         17.975     2.580       0.000     0.230     1.425     0.310     1.615
#> 32         10.135     7.535       0.320     0.000     2.450     0.000     1.805
#> 33         22.120     0.570       0.000     0.000     1.250     0.000     2.020
#> 34         12.115     1.460       0.000     0.000     0.000     1.425     1.995
#> 35         14.290     0.540       0.255     0.000     0.340     0.000     1.030
#> 36          4.085     0.640       0.000     0.000     0.690     0.000     0.280
#> 37          7.530     0.195       0.000     0.000     0.075     0.000     1.245
#> 38         23.755     2.540       0.055     0.270     0.210     0.300     2.530
#> 39         17.300     3.725       0.000     0.000     0.185     0.720     0.560
#> 40         20.160     2.300       0.000     0.000     0.970     0.000     4.310
#> 41         17.580     1.840       0.000     0.325     0.000     0.000     0.750
#> 42         11.910     0.750       0.660     0.000     0.000     0.320     1.755
#> 43         29.235     1.770       0.000     0.000     1.060     0.485     3.145
#> 44         14.145     4.385       0.000     0.010     0.750     0.530     4.555
#> 45         17.910     0.205       0.000     0.000     1.945     2.000     2.000
#> 46          7.305     1.105       0.000     0.050     2.980     1.810     1.070
#> 47         10.555     2.110       0.000     0.345     0.530     0.065     0.400
#> 48          3.120     1.695       0.380     0.000     1.265     0.405     0.890
#> 49          5.955     0.730       0.000     0.000     0.925     1.715     1.150
#> 50         42.055     3.240       0.000     0.165     0.750     2.615     3.715
#>      T_G_REFINED
#> 1          6.475
#> 2          5.625
#> 3          5.110
#> 4         11.190
#> 5          5.665
#> 6         16.585
#> 7          3.065
#> 8          2.505
#> 9          7.345
#> 10         2.365
#> 11         5.980
#> 12         6.300
#> 13         6.400
#> 14         6.130
#> 15         6.115
#> 16         6.275
#> 17         0.545
#> 18         7.690
#> 19         0.385
#> 20         4.885
#> 21         3.790
#> 22         1.090
#> 23         5.065
#> 24         2.255
#> 25         3.400
#> 26         5.480
#> 27         7.565
#> 28         7.885
#> 29         4.685
#> 30         2.065
#> 31         4.970
#> 32        10.530
#> 33         6.410
#> 34         7.460
#> 35         2.200
#> 36         3.090
#> 37         2.605
#> 38         8.155
#> 39         2.695
#> 40        13.040
#> 41         2.220
#> 42         2.635
#> 43         2.240
#> 44         1.645
#> 45         3.860
#> 46         2.125
#> 47         2.730
#> 48         2.545
#> 49         3.445
#> 50         3.750
#>  [ reached getOption("max.print") -- omitted 8229 rows ]

get_fped("2005/2006", "first")

#>       SEQN RIDAGEYR DRSTZ T_F_CITMLB T_PF_EGGS T_F_OTHER T_PF_MPS_TOTAL
#> 1    31127        0     1       0.38      1.15      0.00           1.50
#> 2    31128       11     1       0.00      0.02      1.26           0.99
#> 3    31129       15     1       0.00      0.39      1.01          10.42
#> 4    31131       44     1       0.00      0.40      0.00           0.42
#> 5    31132       70     1       1.67      0.04      2.72           6.83
#> 6    31133       16     1       0.49      0.00      0.00           3.92
#> 7    31134       73     1       0.00      1.50      1.20           3.50
#> 8    31135        0     1       0.50      0.04      0.46           1.21
#> 9    31137       14     1       0.00      0.13      0.00           2.92
#> 10   31138        3     1       0.77      0.00      0.00           2.02
#> 11   31139       18     1       0.00      0.02      0.01           4.49
#> 12   31140       11     1       1.01      0.02      1.03           2.67
#> 13   31141       16     1       0.00      0.85      1.33           2.71
#> 14   31142        7     1       0.00      0.07      1.26           5.39
#> 15   31143       19     1       0.51      0.38      0.00           2.78
#> 16   31144       21     1       0.00      0.00      0.00           0.79
#> 17   31145       10     1       0.10      0.03      0.03           2.68
#> 18   31147        7     1       0.00      0.02      0.00           0.77
#> 19   31148       16     1       0.00      0.02      0.00           3.50
#> 20   31149       85     1       0.00      0.00      0.40           3.43
#> 21   31150       79     1       1.67      0.02      1.27           5.45
#> 22   31151       59     1       0.00      0.04      0.00           1.71
#> 23   31152       27     1       0.00      4.42      0.00           4.51
#> 24   31153       44     1       2.40      0.00      1.39           0.00
#> 25   31154       62     1       0.00      0.21      0.00           2.47
#> 26   31155       38     1       0.00      0.70      0.63          13.81
#> 27   31156       43     1       0.00      0.00      0.00           1.28
#> 28   31157       11     1       0.00      0.09      0.00           3.02
#> 29   31158       71     1       0.00      0.32      1.27           7.06
#> 30   31159        9     1       1.23      0.08      0.05           3.49
#> 31   31160       39     1       0.00      0.08      0.14           5.01
#> 32   31161       10     1       1.11      0.21      0.48           1.31
#> 33   31162       71     1       0.00      1.76      1.88           0.54
#> 34   31163       17     1       0.00      0.00      0.02           2.34
#> 35   31164       54     1       0.00      1.51      0.00           4.95
#> 36   31165        1     1       0.29      0.01      0.48           0.94
#> 37   31166       15     1       0.00      0.00      0.00           3.39
#> 38   31167       37     1       0.00      1.68      2.51           8.30
#> 39   31168        0     1       0.45      0.00      1.37           7.43
#> 40   31169        3     1       0.00      0.15      0.02           4.03
#> 41   31170       14     1       0.00      0.23      0.00           6.65
#> 42   31171       12     1       0.00      0.03      0.00           4.02
#> 43   31172       20     1       0.00      0.00      0.00           1.04
#> 44   31173        6     1       0.00      1.70      0.00           4.20
#> 45   31174        6     1       0.00      0.84      0.48           4.01
#> 46   31175       75     1       0.00      0.00      0.96           2.20
#> 47   31176        3     1       0.00      0.00      0.00           0.08
#> 48   31177       18     1       0.00      0.00      0.00           8.71
#> 49   31178       14     1       0.00      0.00      0.00           5.11
#> 50   31179       71     1       0.00      0.06      0.21           2.15
#>      T_PF_SOY T_PF_SEAFD_HI T_PF_SEAFD_LOW T_PF_NUTSDS T_SOLID_FATS
#> 1        0.00          0.00           0.00        0.00        38.21
#> 2        0.00          0.00           0.00        0.00        14.61
#> 3        0.00          0.00           0.00        0.00        48.68
#> 4        0.00          0.00           0.22        0.93        39.52
#> 5        0.00          0.00           0.00        0.00         7.52
#> 6        0.00          0.00           0.00        0.00        17.42
#> 7        0.00          0.00           0.00        0.00        36.02
#> 8        0.00          0.00           1.21        0.00        22.05
#> 9        0.00          0.00           0.00        0.00        51.04
#> 10       0.00          0.00           0.00        0.00        19.00
#> 11       0.00          0.00           1.49        0.00         0.04
#> 12       0.00          0.00           0.00        0.00        34.76
#> 13       0.00          0.00           0.00        0.00        44.83
#> 14       0.00          0.61           1.83        0.00        19.48
#> 15       0.00          0.00           0.00        0.00        68.27
#> 16       0.00          0.00           0.00        0.00        49.24
#> 17       0.00          0.00           0.00        0.66        31.50
#> 18       0.00          0.00           0.00        0.00        17.58
#> 19       0.00          0.00           0.00        2.38        57.37
#> 20       0.00          0.00           0.00        0.00        28.56
#> 21       0.00          0.00           0.00        0.00         2.22
#> 22       0.00          0.00           0.00        4.15        18.04
#> 23       0.00          0.00           3.24        0.00        44.29
#> 24       2.61          0.00           0.00       14.16        21.17
#> 25       0.00          0.00           0.00        4.29        32.28
#> 26       0.00          0.00           3.60        0.00        37.87
#> 27       0.00          0.00           0.00        0.00        31.03
#> 28       0.00          0.00           0.00        0.00        47.59
#> 29       0.02          0.00           0.00        0.60        79.81
#> 30       0.00          0.00           0.00        0.00        94.26
#> 31       0.00          0.00           0.00        0.00        41.63
#> 32       0.00          0.00           0.00        0.81        38.56
#> 33       0.00          0.00           0.00        0.00        66.27
#> 34       0.00          0.00           0.00        0.00        27.12
#> 35       0.00          0.00           2.38        0.67        95.30
#> 36       0.00          0.00           0.00        0.00        17.86
#> 37       0.00          0.00           0.00        0.00        21.93
#> 38       0.00          0.00           0.00        0.00        74.86
#> 39       0.00          0.00           0.00        0.00        80.10
#> 40       0.00          0.00           0.00        0.00        49.76
#> 41       0.00          0.00           0.00        0.00        38.37
#> 42       0.00          0.00           0.00        0.00        26.98
#> 43       0.00          0.00           0.00        0.00        13.22
#> 44       0.00          0.42           2.25        0.00        12.42
#> 45       0.00          0.00           0.00        2.00        61.01
#> 46       0.00          0.00           0.00        0.00         6.12
#> 47       0.00          0.00           0.00        0.00        14.38
#> 48       0.00          0.00           0.00        0.00        99.33
#> 49       0.00          0.00           0.00        0.00        53.63
#> 50       0.00          0.00           0.00        1.50        58.34
#>      T_ADD_SUGARS T_V_TOTAL T_V_LEGUMES T_V_DRKGR T_F_TOTAL T_G_WHOLE T_D_TOTAL
#> 1            1.91      0.24        0.00      0.00      2.88      0.00      3.59
#> 2           14.02      1.76        0.00      0.00      1.26      0.00      1.84
#> 3          104.20      0.48        0.00      0.00      1.17      0.00      3.27
#> 4           23.38      0.30        0.00      0.00      1.49      0.00      0.76
#> 5            9.11      0.73        0.00      0.00      4.39      1.66      0.51
#> 6           27.51      0.86        0.00      0.00      0.49      0.00      0.37
#> 7           30.64      1.36        0.21      0.00      1.22      0.57      0.27
#> 8            0.17      1.03        0.00      0.00      0.96      0.18      1.26
#> 9           31.95      1.61        0.00      0.00      0.64      0.00      3.21
#> 10          10.96      1.33        0.00      0.00      2.44      0.32      1.22
#> 11           2.38      2.35        0.00      0.00      0.21      0.00      0.00
#> 12           1.72      0.36        0.00      0.00      2.04      0.00      2.60
#> 13          12.62      0.63        0.00      0.46      1.33      1.32      7.39
#> 14          10.84      0.34        0.00      0.00      1.58      0.14      1.55
#> 15           6.28      2.37        0.00      0.09      2.49      1.25      4.92
#> 16          16.55      0.77        0.00      0.00      0.00      0.00      1.71
#> 17          16.91      0.99        0.00      0.00      0.96      0.00      0.59
#> 18           7.19      0.00        0.00      0.00      0.01      0.00      0.61
#> 19          33.49      0.38        0.00      0.00      0.00      0.00      2.92
#> 20          12.28      1.35        0.00      0.00      0.40      0.00      0.66
#> 21           4.07      1.58        0.00      0.00      4.93      3.05      1.33
#> 22          36.29      0.44        0.00      0.00      0.01      0.59      0.77
#> 23          12.13      2.69        0.00      0.24      0.01      0.00      0.74
#> 24          22.78      1.11        0.00      0.00      5.42      0.49      4.25
#> 25          27.84      1.02        0.00      0.91      0.03      0.10      0.09
#> 26           5.03      1.92        0.00      0.00      0.63      0.00      2.12
#> 27          18.01      0.28        0.05      0.00      2.12      0.00      0.07
#> 28           9.40      3.04        0.00      0.00      1.74      0.00      1.77
#> 29          29.82      1.21        0.00      0.17      1.27      0.00      0.54
#> 30           5.76      2.67        0.00      0.00      1.58      2.03      2.04
#> 31          14.76      1.10        0.00      0.00      0.29      2.63      0.54
#> 32          19.69      1.87        0.00      1.18      1.79      0.00      2.42
#> 33           7.30      1.36        0.00      0.00      2.07      1.73      1.59
#> 34          34.61      0.92        0.00      0.00      0.96      0.51      1.89
#> 35          30.17      1.79        0.00      0.00      0.00      0.00      1.54
#> 36           0.30      0.42        0.00      0.07      1.39      0.06      2.25
#> 37          20.36      0.92        0.00      0.00      0.66      1.48      3.15
#> 38           1.81      1.23        0.45      0.00      2.51      0.00      3.53
#> 39          11.34      1.41        1.25      0.00      2.25      0.00      2.13
#> 40           9.41      2.02        0.00      0.49      3.01      0.34      3.09
#> 41          23.03      2.51        0.00      0.00      0.00      0.85      1.11
#> 42          19.25      1.73        0.00      0.00      0.00      0.00      0.02
#> 43          16.54      0.21        0.43      0.00      0.05      0.00      0.83
#> 44           4.20      1.25        0.00      0.00      0.00      0.00      1.57
#> 45          21.88      0.28        0.00      0.04      0.48      0.75      3.48
#> 46           0.00      0.78        0.00      0.00      1.77      1.00      0.08
#> 47           1.45      0.08        0.25      0.00      1.37      0.00      0.99
#> 48          44.48      2.72        0.00      0.00      0.00      2.90      3.31
#> 49          54.39      1.00        0.51      0.00      0.09      2.15      2.00
#> 50          29.54      2.06        0.00      0.00      2.39      2.59      8.23
#>      T_G_REFINED
#> 1           1.51
#> 2           1.17
#> 3           9.15
#> 4           3.19
#> 5           5.14
#> 6           3.63
#> 7           7.24
#> 8           1.61
#> 9           9.60
#> 10          3.50
#> 11          5.42
#> 12          2.27
#> 13         17.57
#> 14          3.45
#> 15          3.91
#> 16         10.53
#> 17         10.85
#> 18          3.82
#> 19          9.01
#> 20          3.51
#> 21          3.59
#> 22          3.97
#> 23          6.24
#> 24          8.52
#> 25          3.81
#> 26         12.65
#> 27          3.72
#> 28          6.62
#> 29          9.46
#> 30          8.68
#> 31          5.70
#> 32          8.96
#> 33          3.01
#> 34          4.26
#> 35          6.04
#> 36          1.39
#> 37          3.57
#> 38          7.42
#> 39          3.44
#> 40          5.88
#> 41          7.36
#> 42          3.44
#> 43          3.98
#> 44          3.75
#> 45          6.41
#> 46          1.69
#> 47          3.04
#> 48         21.12
#> 49          5.11
#> 50          6.46
#>  [ reached getOption("max.print") -- omitted 9119 rows ]